Feat:Adds support for single and multi-line comments (#56)
* Feat:Adds support for single and multi-line comments

Implements single-line (::) and multi-line (:| ... |:)
commenting functionality in the lexer. This allows
users to add comments to their code, which will be
ignored by the lexer.

This improves code readability and maintainability.

* Feat: Adds multiline comment test cases

Implements support for multiline comments using ":| ... |:" delimiters in the splitter.

This change allows for code comments to span multiple lines. Several new test cases are also included.